Fix FAQ server error + favicon 404
- Extraire allFaqKeys dans data/faq-keys.ts (fichier partagé server/client)
  pour éviter l'import client→server qui causait le crash FAQ
- Ajouter favicon.ico au matcher middleware pour éviter les 404
