+++ components/pages/faq-client.tsx
@@ -13,42 +13,7 @@ import {
 } from "@/components/ui/accordion";
 import { Search, ArrowLeft } from "lucide-react";
 import { cn } from "@/lib/utils";
-
-// Toutes les clés de FAQ avec leurs catégories
-export const allFaqKeys = [
-	// General
-	{ key: "whatIsWisetwin", category: "general" },
-	{ key: "howHelps", category: "general" },
-	{ key: "whatIsWiseTrainer", category: "general" },
-	{ key: "whatIsWisePaper", category: "general" },
-	{ key: "support", category: "general" },
-	// Features
-	{ key: "trainingTypes", category: "features" },
-	{ key: "existingContent", category: "features" },
-	{ key: "languages", category: "features" },
-	{ key: "progressTracking", category: "features" },
-	{ key: "trainingPlans", category: "features" },
-	{ key: "certifications", category: "features" },
-	{ key: "reminders", category: "features" },
-	{ key: "guestMode", category: "features" },
-	{ key: "createSimulator", category: "features" },
-	{ key: "scenarioTypes", category: "features" },
-	{ key: "hybridTraining", category: "features" },
-	// Pricing
-	{ key: "pricingModel", category: "pricing" },
-	{ key: "wisetrainerPricing", category: "pricing" },
-	{ key: "economicBenefits", category: "pricing" },
-	{ key: "freeTrial", category: "pricing" },
-	// Technical
-	{ key: "deploymentTime", category: "technical" },
-	{ key: "security", category: "technical" },
-	{ key: "noVrHeadset", category: "technical" },
-	{ key: "installation", category: "technical" },
-	{ key: "hrIntegration", category: "technical" },
-	{ key: "multiSite", category: "technical" },
-	{ key: "technicalRequirements", category: "technical" },
-	{ key: "gdprCompliance", category: "technical" },
-] as const;
+import { allFaqKeys } from "@/data/faq-keys";
 
 type Category = "all" | "general" | "features" | "pricing" | "technical";
