+++ data/faq-keys.ts
@@ -1,0 +1,34 @@
+export const allFaqKeys = [
+	// General
+	{ key: "whatIsWisetwin", category: "general" },
+	{ key: "howHelps", category: "general" },
+	{ key: "whatIsWiseTrainer", category: "general" },
+	{ key: "whatIsWisePaper", category: "general" },
+	{ key: "support", category: "general" },
+	// Features
+	{ key: "trainingTypes", category: "features" },
+	{ key: "existingContent", category: "features" },
+	{ key: "languages", category: "features" },
+	{ key: "progressTracking", category: "features" },
+	{ key: "trainingPlans", category: "features" },
+	{ key: "certifications", category: "features" },
+	{ key: "reminders", category: "features" },
+	{ key: "guestMode", category: "features" },
+	{ key: "createSimulator", category: "features" },
+	{ key: "scenarioTypes", category: "features" },
+	{ key: "hybridTraining", category: "features" },
+	// Pricing
+	{ key: "pricingModel", category: "pricing" },
+	{ key: "wisetrainerPricing", category: "pricing" },
+	{ key: "economicBenefits", category: "pricing" },
+	{ key: "freeTrial", category: "pricing" },
+	// Technical
+	{ key: "deploymentTime", category: "technical" },
+	{ key: "security", category: "technical" },
+	{ key: "noVrHeadset", category: "technical" },
+	{ key: "installation", category: "technical" },
+	{ key: "hrIntegration", category: "technical" },
+	{ key: "multiSite", category: "technical" },
+	{ key: "technicalRequirements", category: "technical" },
+	{ key: "gdprCompliance", category: "technical" },
+] as const;
