+++ middleware.ts
@@ -35,7 +35,7 @@ export const config = {
 		// - API routes
 		// - Static files (/_next, /image, /video, etc.)
 		// - Favicon, robots, sitemap
-		"/((?!api|_next|image|video|storyset|ressources|sitemap\\.xml|robots\\.txt|.*\\..*|icon\\.ico).*)",
+		"/((?!api|_next|image|video|storyset|ressources|sitemap\\.xml|robots\\.txt|.*\\..*|icon\\.ico|favicon\\.ico).*)",
 		"/(fr|en)/:path*",
 	],
 };
